Initial implementation of 'kwic view mode' for token connect
(default_corparch settings should be tested)
