+++ lib/plugins/abstract/token_connect.pyi
@@ -35,7 +35,8 @@ class Response(object):
     heading:unicode
     note:unicode
 
-    def __init__(self, contents:basestring, renderer:str, status:bool, heading:unicode, note:unicode): ...
+    def __init__(self, contents:basestring, renderer:str, status:bool, heading:unicode, note:unicode,
+                 is_kwic_view:bool): ...
 
     def to_dict(self) -> Dict[str, Any]: ...
 
@@ -65,7 +66,7 @@ class AbstractFrontend(object):
 
     def _fetch_localized_prop(self, prop:str, lang:str) -> unicode: ...
 
-    def export_data(self, data:Any, status:bool, lang:str) -> Response: ...
+    def export_data(self, data:Any, status:bool, lang:str, is_kwic_view:bool) -> Response: ...
 
     def get_heading(self, lang:str) -> unicode: ...
