+++ lib/plugins/default_corparch/__init__.py
@@ -583,7 +583,8 @@ class CorpusArchive(AbstractSearchableCorporaArchive):
 
         token_connect_elm = node.find('token_connect')
         if token_connect_elm is not None:
-            ans.token_connect.providers = [p.text for p in token_connect_elm.findall('provider')]
+            ans.token_connect.providers = [(p.text, token_connect_elm.attrib.get('is_kwic_view'), 0)
+                                           for p in token_connect_elm.findall('provider')]
 
         kwic_connect_elm = node.find('kwic_connect')
         if kwic_connect_elm is not None:
