+++ public/files/js/models/concordance/detail.ts
@@ -27,8 +27,9 @@ import {PageModel} from '../../app/main';
 import {ConcLineModel} from './lines';
 import {AudioPlayer} from './media';
 import * as Immutable from 'immutable';
-import RSVP from 'rsvp';
 import { Action, IFullActionControl } from 'kombo';
+import { Observable, of as rxOf } from 'rxjs';
+import { tap, map } from 'rxjs/operators';
 
 /**
  *
@@ -207,7 +208,7 @@ export class ConcDetailModel extends StatefulModel {
                             this.lineIdx,
                             [],
                             action.payload['position']
-                    ).then(
+                    ).subscribe(
                         () => {
                             this.isBusy = false;
                             this.linesModel.setLineFocus(this.lineIdx, true);
@@ -234,14 +235,13 @@ export class ConcDetailModel extends StatefulModel {
                             action.payload['lineIdx'],
                             [],
                             this.expandLeftArgs.size > 1 && this.expandRightArgs.size > 1 ? 'reload' : null
-                    ).then(
+                    ).subscribe(
                         () => {
                             this.isBusy = false;
                             this.linesModel.setLineFocus(action.payload['lineIdx'], true);
                             this.linesModel.emitChange();
                             this.emitChange();
-                        }
-                    ).catch(
+                        },
                         (err) => {
                             this.isBusy = false;
                             this.emitChange();
@@ -255,12 +255,11 @@ export class ConcDetailModel extends StatefulModel {
                         action.payload['kwicLength'],
                         action.payload['lineIdx']
 
-                    ).then(
+                    ).subscribe(
                         () => {
                             this.tokenConnectIsBusy = false;
                             this.emitChange();
-                        }
-                    ).catch(
+                        },
                         (err) => {
                             this.tokenConnectIsBusy = false;
                             this.emitChange();
@@ -278,13 +277,11 @@ export class ConcDetailModel extends StatefulModel {
                         1,
                         action.payload['lineIdx']
 
-                    ).then(
+                    ).subscribe(
                         () => {
                             this.tokenConnectIsBusy = false;
                             this.emitChange();
-                        }
-
-                    ).catch(
+                        },
                         (err) => {
                             this.tokenConnectIsBusy = false;
                             this.emitChange();
@@ -294,7 +291,7 @@ export class ConcDetailModel extends StatefulModel {
 
                 break;
                 case 'CONCORDANCE_SHOW_WHOLE_DOCUMENT':
-                    this.loadWholeDocument().then(
+                    this.loadWholeDocument().subscribe(
                         () => {
                             this.emitChange();
                         },
@@ -315,7 +312,7 @@ export class ConcDetailModel extends StatefulModel {
                             action.payload['tokenNumber'],
                             action.payload['kwicLength'],
                             action.payload['lineIdx'],
-                            this.expandLeftArgs.size > 1 && this.expandRightArgs.size > 1 ? 'reload' : null).then(
+                            this.expandLeftArgs.size > 1 && this.expandRightArgs.size > 1 ? 'reload' : null).subscribe(
                         () => {
                             this.isBusy = false;
                             this.linesModel.setLineFocus(action.payload['lineIdx'], true);
@@ -338,7 +335,7 @@ export class ConcDetailModel extends StatefulModel {
                             this.kwicTokenNum,
                             this.kwicLength,
                             this.lineIdx,
-                            action.payload['position']).then(
+                            action.payload['position']).subscribe(
                         () => {
                             this.isBusy = false;
                             this.linesModel.setLineFocus(this.lineIdx, true);
@@ -373,8 +370,18 @@ export class ConcDetailModel extends StatefulModel {
                             this.isBusy = true;
                             this.emitChange();
                             return this.reloadSpeechDetail();
+
+                        } else {
+                            this.mode = action.payload['value'];
+                            this.expandLeftArgs = Immutable.List<ExpandArgs>();
+                            this.expandRightArgs = Immutable.List<ExpandArgs>();
+                            this.expaningSide = null;
+                            this.concDetail = null;
+                            this.isBusy = true;
+                            this.emitChange();
+                            return rxOf(null);
                         }
-                    })().then(
+                    })().subscribe(
                         () => {
                             this.isBusy = false;
                             this.emitChange();
@@ -592,9 +599,9 @@ export class ConcDetailModel extends StatefulModel {
     /**
      *
      */
-    private loadWholeDocument():RSVP.Promise<any> {
+    private loadWholeDocument():Observable<any> {
 
-        return this.layoutModel.ajax<AjaxResponse.WideCtx>(
+        return this.layoutModel.ajax$<AjaxResponse.WideCtx>(
             'GET',
             this.layoutModel.createActionUrl('structctx'),
             {
@@ -604,20 +611,22 @@ export class ConcDetailModel extends StatefulModel {
             },
             {}
 
-        ).then(
-            (data) => {
-                this.concDetail = data.content;
-                this.wholeDocumentLoaded = true;
-                this.expandLeftArgs = Immutable.List<ExpandArgs>();
-                this.expandRightArgs = Immutable.List<ExpandArgs>();
-            }
+        ).pipe(
+            tap(
+                (data) => {
+                    this.concDetail = data.content;
+                    this.wholeDocumentLoaded = true;
+                    this.expandLeftArgs = Immutable.List<ExpandArgs>();
+                    this.expandRightArgs = Immutable.List<ExpandArgs>();
+                }
+            )
         );
     }
 
     /**
      *
      */
-    private loadSpeechDetail(corpusId:string, tokenNum:number, kwicLength:number, lineIdx:number, expand?:string):RSVP.Promise<any> {
+    private loadSpeechDetail(corpusId:string, tokenNum:number, kwicLength:number, lineIdx:number, expand?:string):Observable<boolean> {
         const structs = this.layoutModel.getConcArgs().getList('structs');
         const args = this.speechAttrs
                 .map(x => `${this.speechOpts.speakerIdAttr[0]}.${x}`)
@@ -636,32 +645,33 @@ export class ConcDetailModel extends StatefulModel {
         return this.loadConcDetail(corpusId, tokenNum, kwicLength, lineIdx, args, expand);
     }
 
-    private reloadSpeechDetail():RSVP.Promise<any> {
+    private reloadSpeechDetail():Observable<boolean> {
         return this.loadSpeechDetail(this.corpusId, this.kwicTokenNum, this.kwicLength, this.lineIdx);
     }
 
-    private loadTokenConnect(corpusId:string, tokenNum:number, numTokens:number, lineIdx:number):RSVP.Promise<boolean> {
+    private loadTokenConnect(corpusId:string, tokenNum:number, numTokens:number, lineIdx:number):Observable<boolean> {
         return (() => {
             if (this.tokenConnectPlg) {
                 return this.tokenConnectPlg.fetchTokenConnect(corpusId, tokenNum, numTokens);
 
             } else {
-                return RSVP.resolve<PluginInterfaces.TokenConnect.TCData>(null);
+                return rxOf<PluginInterfaces.TokenConnect.TCData>(null);
             }
-        })().then(
-            (data) => {
-                if (data) {
-                    this.tokenConnectData = {
-                        token: data.token,
-                        renders: data.renders
-                    };
-                    this.lineIdx = lineIdx;
-                    return true;
-
-                } else {
-                    return false;
+        })().pipe(
+            tap(
+                (data) => {
+                    if (data) {
+                        this.tokenConnectData = {
+                            token: data.token,
+                            renders: data.renders
+                        };
+                        this.lineIdx = lineIdx;
+                    }
                 }
-            }
+            ),
+            map(
+                (data) => data ? true : false
+            )
         );
     }
 
@@ -669,7 +679,7 @@ export class ConcDetailModel extends StatefulModel {
      *
      */
     private loadConcDetail(corpusId:string, tokenNum:number, kwicLength:number, lineIdx:number, structs:Array<string>,
-                expand?:string):RSVP.Promise<any> {
+                expand?:string):Observable<boolean> {
         this.corpusId = corpusId;
         this.kwicTokenNum = tokenNum;
         this.kwicLength = kwicLength;
@@ -713,36 +723,39 @@ export class ConcDetailModel extends StatefulModel {
         this.isBusy = true;
         this.emitChange();
 
-        return this.layoutModel.ajax<AjaxResponse.WideCtx>(
+        return this.layoutModel.ajax$<AjaxResponse.WideCtx>(
             'GET',
             this.layoutModel.createActionUrl('widectx'),
             args,
             {}
 
-        ).then(
-            (data) => {
-                this.concDetail = data.content;
-                if (data.expand_left_args) {
-                    this.expandLeftArgs = this.expandLeftArgs.push([
-                        data.expand_left_args.detail_left_ctx, data.expand_left_args.detail_right_ctx
-                    ]);
+        ).pipe(
+            tap(
+                (data) => {
+                    this.concDetail = data.content;
+                    if (data.expand_left_args) {
+                        this.expandLeftArgs = this.expandLeftArgs.push([
+                            data.expand_left_args.detail_left_ctx, data.expand_left_args.detail_right_ctx
+                        ]);
 
-                } else {
-                    this.expandLeftArgs = this.expandLeftArgs.push(null);
-                }
-                if (data.expand_right_args) {
-                    this.expandRightArgs = this.expandRightArgs.push([
-                        data.expand_right_args.detail_left_ctx, data.expand_right_args.detail_right_ctx
-                    ]);
+                    } else {
+                        this.expandLeftArgs = this.expandLeftArgs.push(null);
+                    }
+                    if (data.expand_right_args) {
+                        this.expandRightArgs = this.expandRightArgs.push([
+                            data.expand_right_args.detail_left_ctx, data.expand_right_args.detail_right_ctx
+                        ]);
 
-                } else {
-                    this.expandRightArgs = this.expandRightArgs.push(null);
+                    } else {
+                        this.expandRightArgs = this.expandRightArgs.push(null);
+                    }
                 }
-            }
+            ),
+            map(d =>  !!d)
         );
     }
 
-    private reloadConcDetail():RSVP.Promise<any> {
+    private reloadConcDetail():Observable<boolean> {
         return this.loadConcDetail(this.corpusId, this.kwicTokenNum, this.kwicLength, this.lineIdx, [], 'reload');
     }
 
@@ -825,7 +838,7 @@ export class RefsDetailModel extends StatefulModel {
                 case 'CONCORDANCE_SHOW_REF_DETAIL':
                     this.isBusy = true;
                     this.emitChange();
-                    this.loadRefs(action.payload['corpusId'], action.payload['tokenNumber'], action.payload['lineIdx']).then(
+                    this.loadRefs(action.payload['corpusId'], action.payload['tokenNumber'], action.payload['lineIdx']).subscribe(
                         () => {
                             this.linesModel.setLineFocus(action.payload['lineIdx'], true);
                             this.linesModel.emitChange();
@@ -870,17 +883,20 @@ export class RefsDetailModel extends StatefulModel {
         }
     }
 
-    private loadRefs(corpusId:string, tokenNum:number, lineIdx:number):RSVP.Promise<any> {
-        return this.layoutModel.ajax<AjaxResponse.FullRef>(
+    private loadRefs(corpusId:string, tokenNum:number, lineIdx:number):Observable<boolean> {
+        return this.layoutModel.ajax$<AjaxResponse.FullRef>(
             'GET',
             this.layoutModel.createActionUrl('fullref'),
             {corpname: corpusId, pos: tokenNum}
 
-        ).then(
-            (data) => {
-                this.lineIdx = lineIdx;
-                this.data = Immutable.List<RefsColumn>(data.Refs);
-            }
+        ).pipe(
+            tap(
+                (data) => {
+                    this.lineIdx = lineIdx;
+                    this.data = Immutable.List<RefsColumn>(data.Refs);
+                }
+            ),
+            map(data => !!data)
         );
     }
